Сode optimization
When looking through the code, I realized that I did not divide the functions into classes and did not use their capabilities. Therefore, classes were added: obstacles, apple, stone and snake.
I also rewrote the code for updating the leaderboard through a loop, for simplicity.
I would also like to implement a change in the speed of the snake during the game, but I don’t know how to do it yet.
Please tell me who knows how to implement it.